speed up debugging
