+++ pytest.ini
@@ -1,6 +1,6 @@
 [pytest]
 minversion = 4.5
-testpaths = tests/epyccel/recognised_functions
+testpaths = tests
 addopts = --strict-markers
 markers =
     parallel: test to be run using 'mpiexec'
