+++ tests/a_recognised_functions/test_math_funcs.py
